+++ src/main/java/com/technologygarden/service/impl/DegreeServiceImpl.java
@@ -3,8 +3,13 @@ package com.technologygarden.service.impl;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.technologygarden.dao.DegreeMapper;
+import com.technologygarden.dao.EmployeeMapper;
+import com.technologygarden.dao.LegalPersonMapper;
 import com.technologygarden.entity.Degree;
+import com.technologygarden.entity.Employee;
+import com.technologygarden.entity.LegalPerson;
 import com.technologygarden.entity.ResultBean.ResultBean;
+import com.technologygarden.entity.ResultBean.ResultStatus;
 import com.technologygarden.service.DegreeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,9 +19,13 @@ import java.util.List;
 @Service("DegreeService")
 public class DegreeServiceImpl implements DegreeService {
     private final DegreeMapper degreeMapper;
+    private final EmployeeMapper employeeMapper;
+    private final LegalPersonMapper legalPersonMapper;
     @Autowired
-    public DegreeServiceImpl(DegreeMapper degreeMapper) {
+    public DegreeServiceImpl(DegreeMapper degreeMapper, EmployeeMapper employeeMapper, LegalPersonMapper legalPersonMapper) {
         this.degreeMapper = degreeMapper;
+        this.employeeMapper = employeeMapper;
+        this.legalPersonMapper = legalPersonMapper;
     }
 
     @Override
@@ -38,6 +47,12 @@ public class DegreeServiceImpl implements DegreeService {
 
     @Override
     public ResultBean<?> deleteDegree(Integer id) {
+        List<Employee> employeeList =employeeMapper.selectByxId(id);
+        List<LegalPerson> legalPeople =legalPersonMapper.selectLegalPersonByDegreeId(id);
+        //判断有无员工与学位关联
+        if(employeeList.size()>0 || legalPeople.size()>0){
+            return new ResultBean<>(ResultStatus.DELETE_ERROR.getCode(), ResultStatus.DELETE_ERROR.getMessage());
+        }
         return new ResultBean<>(degreeMapper.deleteByPrimaryKey(id));
     }
